lets use u16

+++ src/consts.rs
@@ -1,4 +1,4 @@
 use std::error::Error;
 
-pub type Number = i32;
+pub type Number = u16;
 pub type DynResult<T> = Result<T, Box<dyn Error>>;
